+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+* Add example of reading large response body.
+
+--------------------------------------------------------------------------------
+
 Version 316:
 
 * Disable GHA CI for clang-9.